practical_django/requirements.txt
appnope==0.1.0
backcall==0.1.0
cairocffi==1.0.2
CairoSVG==2.3.1
cffi==1.12.3
cssselect2==0.2.1
decorator==4.4.0
defusedxml==0.6.0
Django==2.2
django-debug-toolbar==1.11
django-extensions==2.1.6
django-filter==2.1.0
django-tables2==2.0.6
django-webpack-loader==0.6.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.3
factory-boy==2.11.1
Faker==1.0.5
html5lib==1.0.1
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.9
psycopg2==2.8.2
ptyprocess==0.6.0
pycparser==2.19
pydotplus==2.0.2
Pygments==2.3.1
pyparsing==2.4.0
Pyphen==0.9.5
python-dateutil==2.8.0
pytz==2019.1
selenium==3.141.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
tinycss2==1.0.2
traitlets==4.3.2
urllib3==1.25.2
wcwidth==0.1.7
WeasyPrint==47
webencodings==0.5.1
Werkzeug==0.15.2
